Fix argcomplete
It must be called unconditionally when starting script without
arguments. It was not working before, because in case of empty argument
list, print_usage was called and script exited. As a result argcomplete
code was never reached.

The way argcomplete works is it exports a bunch of bash-specific
environment variables and also few argcomplete-specific ones, then calls
script without arguments and does its stuff in autocomplete method of
argcomplete module. If this method is not called when calling script
without arguments, it simply does not work.

As a debugging facility for the future, add debug string on argcomplete
call.