Add simple argv parser for early log level setting
If argparse is to be used for parsing log level in order to enable
debug, log level will be set very far into main menu processing, as it
is mostly argument parsing and not much else.

To be able to see all the messages from the very beginning, parse -v and
--verbose manually, with error handling hidden, so that improper
argument will not be visible in this primitive parser. Anyway, log is
also set later after processing, so user should still be able to see
what mistake he made.